+++ Framework/Extensions/SpriteBatchExtensions.cs
@@ -16,7 +16,8 @@
         /// <param name="transform">The transform.</param>
         /// <param name="rotationAngle">The rotation angle.</param>
         /// <param name="color">The color.</param>
-        public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Transform transform, float rotationAngle, Color color) {
+        /// <param name="orientation">The orientation.</param>
+        public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Transform transform, float rotationAngle, Color color, SpriteEffects orientation) {
             if (sprite?.Texture != null && transform != null) {
                 spriteBatch.Draw(
                     sprite.Texture,
@@ -26,7 +27,7 @@
                     rotationAngle,
                     Vector2.Zero,
                     transform.Scale,
-                    SpriteEffects.FlipVertically,
+                    orientation,
                     0f);
             }
         }
@@ -40,7 +41,8 @@
         /// <param name="scale">The scale.</param>
         /// <param name="rotationAngle">The rotation angle.</param>
         /// <param name="color">The color.</param>
-        public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Vector2 position, Vector2 scale, float rotationAngle, Color color) {
+        /// <param name="orientation">The orientation.</param>
+        public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Vector2 position, Vector2 scale, float rotationAngle, Color color, SpriteEffects orientation) {
             if (sprite?.Texture != null) {
                 spriteBatch.Draw(
                     sprite.Texture,
@@ -50,7 +52,7 @@
                     rotationAngle,
                     Vector2.Zero,
                     scale,
-                    SpriteEffects.FlipVertically,
+                    orientation,
                     0f);
             }
         }
@@ -64,7 +66,7 @@
         /// <param name="scale">The scale.</param>
         /// <param name="color">The color.</param>
         public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Vector2 position, Vector2 scale, Color color) {
-            spriteBatch.Draw(sprite, position, scale, 0f, color);
+            spriteBatch.Draw(sprite, position, scale, 0f, color, SpriteEffects.FlipVertically);
         }
 
         /// <summary>
@@ -75,7 +77,19 @@
         /// <param name="transform">The transform.</param>
         /// <param name="color">The color.</param>
         public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, RotatableTransform transform, Color color) {
-            spriteBatch.Draw(sprite, transform, transform.Rotation.Angle, color);
+            spriteBatch.Draw(sprite, transform, transform.Rotation.Angle, color, SpriteEffects.FlipVertically);
+        }
+
+        /// <summary>
+        /// Draws the specified sprite.
+        /// </summary>
+        /// <param name="spriteBatch">The sprite batch.</param>
+        /// <param name="sprite">The sprite.</param>
+        /// <param name="transform">The transform.</param>
+        /// <param name="color">The color.</param>
+        /// <param name="orientation">The orientation.</param>
+        public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, RotatableTransform transform, Color color, SpriteEffects orientation) {
+            spriteBatch.Draw(sprite, transform, transform.Rotation.Angle, color, orientation);
         }
 
         /// <summary>
@@ -86,7 +100,19 @@
         /// <param name="transform">The transform.</param>
         /// <param name="color">The color.</param>
         public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Transform transform, Color color) {
-            spriteBatch.Draw(sprite, transform, 0f, color);
+            spriteBatch.Draw(sprite, transform, 0f, color, SpriteEffects.FlipVertically);
+        }
+
+        /// <summary>
+        /// Draws the specified sprite.
+        /// </summary>
+        /// <param name="spriteBatch">The sprite batch.</param>
+        /// <param name="sprite">The sprite.</param>
+        /// <param name="transform">The transform.</param>
+        /// <param name="color">The color.</param>
+        /// <param name="orientation">The orientation.</param>
+        public static void Draw(this SpriteBatch spriteBatch, Sprite sprite, Transform transform, Color color, SpriteEffects orientation) {
+            spriteBatch.Draw(sprite, transform, 0f, color, orientation);
         }
 
         /// <summary>
@@ -98,7 +124,7 @@
         /// <param name="scale">The scale.</param>
         /// <param name="rotationAngle">The rotation angle.</param>
         /// <param name="color">The color.</param>
-        public static void Draw(this SpriteBatch spriteBatch, Font font, string text, Vector2 position, Vector2 scale, float rotationAngle, Color color) {
+        public static void Draw(this SpriteBatch spriteBatch, Font font, string text, Vector2 position, Vector2 scale, float rotationAngle, Color color, SpriteEffects orientation) {
             if (font?.SpriteFont != null) {
                 spriteBatch.DrawString(
                   font.SpriteFont,
@@ -108,9 +134,35 @@
                   rotationAngle,
                   Vector2.Zero,
                   scale,
-                  SpriteEffects.FlipVertically,
+                  orientation,
                   0f);
             }
+        }
+
+        /// <summary>
+        /// Draws the text with the provided font.
+        /// </summary>
+        /// <param name="spriteBatch">The sprite batch.</param>
+        /// <param name="font">The font.</param>
+        /// <param name="text">The text.</param>
+        /// <param name="rotatableTransform">The rotatable transform.</param>
+        /// <param name="color">The color.</param>
+        /// <param name="orientation">The orientation.</param>
+        public static void Draw(this SpriteBatch spriteBatch, Font font, string text, RotatableTransform rotatableTransform, Color color, SpriteEffects orientation) {
+            spriteBatch.Draw(font, text, rotatableTransform.Position, rotatableTransform.Scale, rotatableTransform.Rotation.Angle, color, orientation);
+        }
+
+        /// <summary>
+        /// Draws the text with the provided font.
+        /// </summary>
+        /// <param name="spriteBatch">The sprite batch.</param>
+        /// <param name="font">The font.</param>
+        /// <param name="text">The text.</param>
+        /// <param name="transform">The transform.</param>
+        /// <param name="color">The color.</param>
+        /// <param name="orientation">The orientation.</param>
+        public static void Draw(this SpriteBatch spriteBatch, Font font, string text, Transform transform, Color color, SpriteEffects orientation) {
+            spriteBatch.Draw(font, text, transform.Position, transform.Scale, 0f, color, orientation);
         }
 
         /// <summary>
@@ -122,7 +174,7 @@
         /// <param name="transform">The transform.</param>
         /// <param name="color">The color.</param>
         public static void Draw(this SpriteBatch spriteBatch, Font font, string text, RotatableTransform rotatableTransform, Color color) {
-            spriteBatch.Draw(font, text, rotatableTransform.Position, rotatableTransform.Scale, rotatableTransform.Rotation.Angle, color);
+            spriteBatch.Draw(font, text, rotatableTransform.Position, rotatableTransform.Scale, rotatableTransform.Rotation.Angle, color, SpriteEffects.FlipVertically);
         }
 
         /// <summary>
@@ -134,7 +186,7 @@
         /// <param name="transform">The transform.</param>
         /// <param name="color">The color.</param>
         public static void Draw(this SpriteBatch spriteBatch, Font font, string text, Transform transform, Color color) {
-            spriteBatch.Draw(font, text, transform.Position, transform.Scale, 0f, color);
+            spriteBatch.Draw(font, text, transform.Position, transform.Scale, 0f, color, SpriteEffects.FlipVertically);
         }
     }
 }